Update Google Fonts instructions to cope with UI updates
https://fonts.google.com/ seems to have been updated (at least a
year ago, according to the Wayback Machine) and now has a fairly
different flow to get to the information you want about how to use
the fonts - ie there's no more 'Quick-use' button, as described in
the previous instructions.

I spent ages looking for the 'Quick-use' when trying to follow this
lesson plan, until eventually realising I just needed to click on the
'plus' icon!